+++ Lilu/Headers/kern_devinfo.hpp
@@ -168,9 +168,28 @@ private:
 
 	/**
 	 *  The IGPU property to force-disable any external GPU if found.
-	 *  For user configuration only! Use requestedExternalSwitchOff!
+	 *  For user configuration only! Use processSwitchOff()!
 	 */
 	static constexpr const char *RequestedExternalSwitchOffName {"disable-external-gpu"};
+
+	/**
+	 *  The GPU property to force-disable any this external GPU.
+	 *  For user configuration only! Use processSwitchOff()!
+	 */
+	static constexpr const char *RequestedGpuSwitchOffName {"disable-gpu"};
+
+	/**
+	 *  The GPU property to force-disable any this external GPU with minimum kernel version (inclusive).
+	 *  For user configuration only! Use processSwitchOff()!
+	 */
+	static constexpr const char *RequestedGpuSwitchOffMinKernelName {"disable-gpu-min"};
+
+	/**
+	 *  The GPU property to force-disable any this external GPU with maximum kernel version (inclusive).
+	 *  For user configuration only! Use processSwitchOff()!
+	 */
+	static constexpr const char *RequestedGpuSwitchOffMaxKernelName {"disable-gpu-max"};
+
 
 	/**
 	 *  Known platform ids used by Intel GPU kexts
@@ -282,6 +301,11 @@ public:
 	 *  @param d  device list
 	 */
 	EXPORT static void deleter(DeviceInfo *d NONNULL);
+
+	/**
+	 *  Perform device switch-off as prescribed by the properties injected.
+	 */
+	EXPORT void processSwitchOff();
 };
 
 /**